+++ MOSBackend/MOS.Domain/Entities/Resumes/ResumeSkill.cs
@@ -4,7 +4,7 @@ namespace MOS.Domain.Entities.Resumes;
 
 public class ResumeSkill : Entity<int>
 {
-    public long ResumeId { get; set; }
+    public int ResumeId { get; set; }
     public string Name { get; set; } = "";
     public SkillLevelType Level { get; set; }
 }
